# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aegis API",
    "description": "Orquestador del agente financiero: propuestas, políticas, Guardian y auditoría.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/policy": {
      "get": {
        "summary": "Reglas y modo de operación actuales",
        "tags": [
          "policy"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {
                      "type": "object",
                      "properties": {
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "5"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "20"
                        },
                        "requireConfirmationForNewDestination": {
                          "type": "boolean",
                          "default": true
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          },
                          "minItems": 1,
                          "default": [
                            "XLM",
                            "USDC_TEST"
                          ]
                        },
                        "maxOperationsPerHour": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "10"
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ],
                          "default": "MANUAL"
                        },
                        "paused": {
                          "type": "boolean",
                          "default": false
                        },
                        "proposalTtlMinutes": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        }
                      },
                      "additionalProperties": false
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ]
                        },
                        "paused": {
                          "type": "boolean"
                        },
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          }
                        },
                        "remainingDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        }
                      },
                      "required": [
                        "mode",
                        "paused",
                        "maxAmountPerOperation",
                        "maxDailyAmount",
                        "minimumReserve",
                        "allowedAssets",
                        "remainingDailyAmount"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "config",
                    "summary"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Actualiza las reglas (parcial)",
        "tags": [
          "policy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxAmountPerOperation": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,7})?$",
                    "default": "5"
                  },
                  "maxDailyAmount": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,7})?$",
                    "default": "20"
                  },
                  "requireConfirmationForNewDestination": {
                    "type": "boolean",
                    "default": true
                  },
                  "allowedAssets": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "XLM",
                        "USDC_TEST"
                      ]
                    },
                    "minItems": 1,
                    "default": [
                      "XLM",
                      "USDC_TEST"
                    ]
                  },
                  "maxOperationsPerHour": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "default": 10
                  },
                  "minimumReserve": {
                    "type": "string",
                    "pattern": "^\\d+(\\.\\d{1,7})?$",
                    "default": "10"
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "MANUAL",
                      "AUTONOMOUS"
                    ],
                    "default": "MANUAL"
                  },
                  "paused": {
                    "type": "boolean",
                    "default": false
                  },
                  "proposalTtlMinutes": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "minimum": 0,
                    "default": 10
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {
                      "type": "object",
                      "properties": {
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "5"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "20"
                        },
                        "requireConfirmationForNewDestination": {
                          "type": "boolean",
                          "default": true
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          },
                          "minItems": 1,
                          "default": [
                            "XLM",
                            "USDC_TEST"
                          ]
                        },
                        "maxOperationsPerHour": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "10"
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ],
                          "default": "MANUAL"
                        },
                        "paused": {
                          "type": "boolean",
                          "default": false
                        },
                        "proposalTtlMinutes": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        }
                      },
                      "additionalProperties": false
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ]
                        },
                        "paused": {
                          "type": "boolean"
                        },
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          }
                        },
                        "remainingDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        }
                      },
                      "required": [
                        "mode",
                        "paused",
                        "maxAmountPerOperation",
                        "maxDailyAmount",
                        "minimumReserve",
                        "allowedAssets",
                        "remainingDailyAmount"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "config",
                    "summary"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/policy/pause": {
      "post": {
        "summary": "Pausa o reactiva al agente",
        "tags": [
          "policy"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paused": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "paused"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config": {
                      "type": "object",
                      "properties": {
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "5"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "20"
                        },
                        "requireConfirmationForNewDestination": {
                          "type": "boolean",
                          "default": true
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          },
                          "minItems": 1,
                          "default": [
                            "XLM",
                            "USDC_TEST"
                          ]
                        },
                        "maxOperationsPerHour": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$",
                          "default": "10"
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ],
                          "default": "MANUAL"
                        },
                        "paused": {
                          "type": "boolean",
                          "default": false
                        },
                        "proposalTtlMinutes": {
                          "type": "integer",
                          "exclusiveMinimum": true,
                          "minimum": 0,
                          "default": 10
                        }
                      },
                      "additionalProperties": false
                    },
                    "summary": {
                      "type": "object",
                      "properties": {
                        "mode": {
                          "type": "string",
                          "enum": [
                            "MANUAL",
                            "AUTONOMOUS"
                          ]
                        },
                        "paused": {
                          "type": "boolean"
                        },
                        "maxAmountPerOperation": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "maxDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "minimumReserve": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        },
                        "allowedAssets": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "XLM",
                              "USDC_TEST"
                            ]
                          }
                        },
                        "remainingDailyAmount": {
                          "type": "string",
                          "pattern": "^\\d+(\\.\\d{1,7})?$"
                        }
                      },
                      "required": [
                        "mode",
                        "paused",
                        "maxAmountPerOperation",
                        "maxDailyAmount",
                        "minimumReserve",
                        "allowedAssets",
                        "remainingDailyAmount"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "config",
                    "summary"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/audit": {
      "get": {
        "summary": "Bitácora de auditoría del usuario, con el estado de la cadena de hashes",
        "tags": [
          "policy"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "minLength": 1
                          },
                          "userId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "proposalId": {
                            "type": "string",
                            "minLength": 1,
                            "nullable": true
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "PROPOSAL_CREATED",
                              "POLICY_EVALUATED",
                              "GUARDIAN_EVALUATED",
                              "STATUS_CHANGED",
                              "USER_APPROVED",
                              "USER_REJECTED",
                              "AGENT_SIGNED",
                              "TX_SUBMITTED",
                              "TX_CONFIRMED",
                              "TX_FAILED",
                              "POLICY_UPDATED",
                              "KILL_SWITCH_TOGGLED",
                              "DESTINATION_CREATED",
                              "DESTINATION_UPDATED",
                              "PROPOSAL_EXPIRED",
                              "PROPOSAL_ABANDONED",
                              "AUTH_LOGIN"
                            ]
                          },
                          "payload": {
                            "type": "object",
                            "additionalProperties": {}
                          },
                          "previousHash": {
                            "type": "string",
                            "nullable": true
                          },
                          "hash": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "type",
                          "payload",
                          "previousHash",
                          "hash",
                          "createdAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "chain": {
                      "type": "object",
                      "properties": {
                        "valid": {
                          "type": "boolean"
                        },
                        "brokenAt": {
                          "type": "string"
                        },
                        "verifiedEvents": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "complete": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "valid",
                        "verifiedEvents",
                        "complete"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "events",
                    "chain"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}